Added interaction with the Postgresql database, plus unit and integration tests for the application. Handled the database schema with Flyway. Dealt with image generation, Docker, Cloud Native Buildpacks, and vulnerability scanning.

+++ src/main/java/com/crunchhub/catalogservice/domain/GameAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package com.crunchhub.catalogservice.domain;
 
 public class GameAlreadyExistsException extends RuntimeException {
-    public GameAlreadyExistsException(int id) {
+    public GameAlreadyExistsException(Long id) {
         super("Game with id " + id + " already exists");
     }
 }
