+++ src/main/java/com/crunchhub/catalogservice/domain/GameNotFoundException.java
@@ -1,7 +1,7 @@
 package com.crunchhub.catalogservice.domain;
 
 public class GameNotFoundException extends RuntimeException {
-    public GameNotFoundException(int id) {
+    public GameNotFoundException(Long id) {
         super("Game with id " + id + " not found");
     }
 }
